refactor: fixes and renames
